refactor(team): extract OpenCode runtime delivery relay UI helpers from the teams IPC module
The teams IPC module carried the whole OpenCode runtime delivery relay UI race
alongside the handler registrations: the UI timeout constants, the relay result
type aliases, the wait/enrich/status-to-result helpers and a private
promise-timeout utility. None of it is IPC wiring, and only one handler
(handleSendMessage, the isOpenCodeRecipient branch) uses it.

Move the generic promise-timeout helper to src/main/ipc/teams/withTimeoutValue.ts
and the relay UI race to src/main/ipc/teams/openCodeRuntimeDeliveryRelayUi.ts.
Only waitForOpenCodeRuntimeRelayForUi is exported from the relay module; the
constants, the OpenCodeMemberInboxRelayResult / OpenCodeMemberInboxDelivery
aliases and the five sub-helpers stay module-private, exactly as private as they
were before. The now-unused TeamOpenCodeMemberInboxRelayResult type import is
dropped from teams.ts and re-declared in the new module against the same
relative contracts path teams.ts used.

Behaviour-free code motion: the moved code is byte-identical to its previous
form apart from the two added export keywords (verified by diffing the new
modules against the corresponding slices of the original file). teams.ts gains
only two import lines. No test was changed - the existing 226 tests in
test/main/ipc/teams.test.ts, including the 6s/1s fake-timer UI-timeout suite
that drives this code through the sendMessage handler, pass unmodified.

teams.ts: 5661 -> 5430 lines.
